+++ R/client_endpoint.R
@@ -6,19 +6,19 @@
 #' @param key The access key for the storage account.
 #' @param token An Azure Active Directory (AAD) authentication token. This can be either a string, or an object of class AzureToken created by [AzureRMR::get_azure_token]. The latter is the recommended way of doing it, as it allows for automatic refreshing of expired tokens.
 #' @param sas A shared access signature (SAS) for the account.
-#' @param api_version The storage API version to use when interacting with the host. Defaults to `"2018-06-17"` for the ADLSgen2 endpoint, and `"2018-03-28"` for the others.
+#' @param api_version The storage API version to use when interacting with the host. Defaults to `"2018-11-09"`.
 #' @param x For the print method, a storage endpoint object.
 #' @param ... For the print method, further arguments passed to lower-level functions.
 #'
 #' @details
 #' This is the starting point for the client-side storage interface in AzureRMR. `storage_endpoint` is a generic function to create an endpoint for any type of Azure storage while `adls_endpoint`, `blob_endpoint` and `file_endpoint` create endpoints for those types.
 #'
-#' If multiple authentication objects are supplied, they are used in this order of priority: first an access key, then an AAD token, then a SAS. If no authentication objects are supplied, only public (anonymous) access to the endpoint is possible. Note that authentication with a SAS is not currently supported by ADLSgen2.
+#' If multiple authentication objects are supplied, they are used in this order of priority: first an access key, then an AAD token, then a SAS. If no authentication objects are supplied, only public (anonymous) access to the endpoint is possible.
 #'
 #' @return
 #' `storage_endpoint` returns an object of S3 class `"adls_endpoint"`, `"blob_endpoint"`, `"file_endpoint"`, `"queue_endpoint"` or `"table_endpoint"` depending on the type of endpoint. All of these also inherit from class `"storage_endpoint"`. `adls_endpoint`, `blob_endpoint` and `file_endpoint` return an object of the respective class.
 #'
-#' Currently AzureStor only includes methods for interacting with ADLSgen2, blob and file storage.
+#' Note that while endpoint classes exist for all storage types, currently AzureStor only includes methods for interacting with ADLSgen2, blob and file storage.
 #'
 #' @seealso
 #' [create_storage_account], [adls_filesystem], [create_adls_filesystem], [file_share], [create_file_share], [blob_container], [create_blob_container]
@@ -60,9 +60,6 @@ storage_endpoint <- function(endpoint, key=NULL, token=NULL, sas=NULL, api_versi
         else getOption("azure_storage_api_version")
     }
 
-    if(type == "adls" && !is_empty(sas))
-        warning("ADLSgen2 does not support authentication with a shared access signature")
-
     obj <- list(url=endpoint, key=key, token=token, sas=sas, api_version=api_version)
     class(obj) <- c(paste0(type, "_endpoint"), "storage_endpoint")
     obj
@@ -74,7 +71,7 @@ blob_endpoint <- function(endpoint, key=NULL, token=NULL, sas=NULL,
                           api_version=getOption("azure_storage_api_version"))
 {
     if(!is_endpoint_url(endpoint, "blob"))
-        stop("Not a blob endpoint", call.=FALSE)
+        warning("Not a recognised blob endpoint", call.=FALSE)
 
     obj <- list(url=endpoint, key=key, token=token, sas=sas, api_version=api_version)
     class(obj) <- c("blob_endpoint", "storage_endpoint")
@@ -87,7 +84,7 @@ file_endpoint <- function(endpoint, key=NULL, token=NULL, sas=NULL,
                           api_version=getOption("azure_storage_api_version"))
 {
     if(!is_endpoint_url(endpoint, "file"))
-        stop("Not a file endpoint", call.=FALSE)
+        warning("Not a recognised file endpoint", call.=FALSE)
 
     obj <- list(url=endpoint, key=key, token=token, sas=sas, api_version=api_version)
     class(obj) <- c("file_endpoint", "storage_endpoint")
@@ -100,10 +97,7 @@ adls_endpoint <- function(endpoint, key=NULL, token=NULL, sas=NULL,
                           api_version=getOption("azure_adls_api_version"))
 {
     if(!is_endpoint_url(endpoint, "adls"))
-        stop("Not an ADLS Gen2 endpoint", call.=FALSE)
-
-    if(!is_empty(sas))
-        warning("ADLSgen2 does not support authentication with a shared access signature")
+        warning("Not a recognised ADLS Gen2 endpoint", call.=FALSE)
 
     obj <- list(url=endpoint, key=key, token=token, sas=sas, api_version=api_version)
     class(obj) <- c("adls_endpoint", "storage_endpoint")