+++ tests/testthat/test03_file.R
@@ -139,27 +139,6 @@ test_that("File client interface works",
     iris3 <- as.data.frame(jsonlite::fromJSON(con))
     expect_identical(iris, iris3)
 
-    # multiple file transfers
-    files <- lapply(1:10, function(f) paste0(sample(letters, 1000, replace=TRUE), collapse=""))
-    filenames <- sapply(1:10, function(n) file.path(tempdir(), sprintf("multitransfer_%d", n)))
-    suppressWarnings(file.remove(filenames))
-    mapply(writeLines, files, filenames)
-
-    create_azure_dir(share, "multi")
-    multiupload_azure_file(share, file.path(tempdir(), "multitransfer_*"), "multi")
-
-    dest_dir <- file.path(tempdir(), "file_multitransfer")
-    suppressWarnings(unlink(dest_dir, recursive=TRUE))
-    dir.create(dest_dir)
-    multidownload_azure_file(share, "multi/multitransfer_*", dest_dir, overwrite=TRUE)
-
-    expect_true(all(sapply(filenames, function(f)
-    {
-        src <- readBin(f, "raw", n=1e5)
-        dest <- readBin(file.path(dest_dir, basename(f)), "raw", n=1e5)
-        identical(src, dest)
-    })))
-
     # ways of deleting a share
     delete_file_share(share, confirm=FALSE)
     delete_file_share(fl, "newshare2", confirm=FALSE)
@@ -266,6 +245,20 @@ test_that("chunked downloading works",
 
     con <- download_azure_file(share, "iris.csv", NULL, blocksize=150)
     expect_identical(readBin(orig_file, "raw", n=1e5), readBin(con, "raw", n=1e5))
+})
+
+
+test_that("Default destination works",
+{
+    fl <- stor$get_file_endpoint()
+    cont <- create_file_share(fl, "defaultdest")
+
+    orig_file <- "../resources/iris.csv"
+    upload_azure_file(cont, orig_file)
+    download_azure_file(cont, basename(orig_file))
+
+    expect_true(file.exists(basename(orig_file)))
+    file.remove(basename(orig_file))
 })